Add the emoji compas submodule
Since this repository is just an iOS wrapper for The Emoji Compass we
need to actually include a direct reference to it.  We don't want to
track changes to the compass itself in here, but we do want to easily be
able to re-build and load the latest versions of it.

Submodules are a good way to accomplish exactly this.

This commit adds the submodule reference.

Issue #1